rtpengine: cast to uint64_t when using time_t value

+++ src/modules/rtpengine/rtpengine.c
@@ -41,6 +41,7 @@
 #include <poll.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <string.h>
 #include <unistd.h>
 #include <ifaddrs.h>
@@ -1336,9 +1337,9 @@ static void rtpengine_rpc_reload(rpc_t* rpc, void* ctx)
 
 	_rtpe_list_version->vernum += 1;
 	_rtpe_list_version->vertime = time(NULL);
-	LM_DBG("current rtpengines list version: %d (%u)\n",
+	LM_DBG("current rtpengines list version: %d (%llu)\n",
 			_rtpe_list_version->vernum,
-			(unsigned int)_rtpe_list_version->vertime);
+			(uint64_t)_rtpe_list_version->vertime);
 	rpc->rpl_printf(ctx, "Ok. Reload successful.");
 }
 
@@ -1856,9 +1857,9 @@ static int build_rtpp_socks(int lmode, int rtest) {
 
 	if(_rtpe_list_vernum_local == _rtpe_list_version->vernum) {
 		/* same version for the list of rtpengines */
-		LM_DBG("same rtpengines list version: %d (%u)\n",
+		LM_DBG("same rtpengines list version: %d (%llu)\n",
 			_rtpe_list_version->vernum,
-			(unsigned int)_rtpe_list_version->vertime);
+			(uint64_t)_rtpe_list_version->vertime);
 		return 0;
 	}
